+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-schemas.js
@@ -27,7 +27,6 @@
         external: this.undefined,
         method: this.undefined,
         link_domain: this.undefined,
-        link_path_parts: this.undefined,
         tool_name: this.undefined,
         percent_scrolled: this.undefined,
         video_current_time: this.undefined,